Rename CustomField to Field and restructure components hieranchy
