+++ src/components/index.js
@@ -1,11 +1,11 @@
 import Vue from 'vue'
 
-import CustomField from '@/components/atoms/CustomField'
+import Field from '@/components/atoms/Field'
 import FeedbackLink from '@/components/atoms/FeedbackLink'
 import RoundedButton from '@/components/atoms/RoundedButton'
 
 export default () => {
-  Vue.component('CustomField', CustomField)
+  Vue.component('Field', Field)
   Vue.component('FeedbackLink', FeedbackLink)
   Vue.component('RoundedButton', RoundedButton)
 }
